server/requirements.txt
openenv[core]>=0.1.13
fastapi>=0.115.0
uvicorn>=0.24.0
numpy>=2.0.0
pydantic>=2.0.0
maptlotlib
jinja2
python-dotenv
openai>=1.0.0






